chagalllte: Add reversed RIL_InitialAttachApn fields
These are needed to fix RIL crashing when trying to attach the initial
APN.

For LTE to work, these fields must be initialized to sane values.
Defaults that I found are "IP" for roamingProtocol and 0 for imsType.

Our custom libril in hardware/samsung will be updated to account for these
requirements.

Change-Id: Iad3c078bf6205d0d45ea5006070f2e1de5a972ea